Switch create_nas_folder method to None
Why these changes are being introduced:
In the last PR, an issue was raised with the create_nas_folder
returning a Boolean instead of raising exceptions in the method itself.

How this addresses that need:
* Update the create_nas_folder() method in the Archive class to return
None by switching the return statements to raise statements instead
* Update the CLI commands that use this method to remove any Boolean
references
* Update the tests for this method to look for expections instead of
Boolean values

Side effects of this change:
None.

Resolves: #9 (GH issue)