Add Dropbox-style SHA256 to Metadata File
Why these changes are being introduced:
Duing testing, Joe asked that the Dropbox-style SHA256 checksum be made
available to the user running the script. We agreed that it was best if
that metadata was just another line in the metadata JSON file that is
saved to the folder on the NAS.

How this addresses that need:
* Rename the static method that calculates the Dropbox-style SHA256
checksum on the file in the NAS (to make it more clear what that method
is actually doing)
* Update the return value for the copy from Dropbox to NAS method to
return a pair of strings (the timestamp and the content_hash metadata)
* Update the bulk_file_copy command to add the content_has metadata to
the JSON file in the target

Side effects of this change:
None.

Related Jira Tickets:
* https://mitlibraries.atlassian.net/browse/IR-217