Update Renovate Config
Why these changes are being introduced:
Still working through what should be the correct configuration for
Renovate. This is an attempt to extend the matching update types to
all the different categories.

How this addresses that need:
* Move the renovate.json into the .github directory (which feels like
a better place for it and is supported by Renovate)
* Add a packageRules stanza to for many different UpdateTypes

Side effects of this change:
None.