Bug Fix: Same Filename Different Extensions
Why these changes are being introduced:
During the first full test with stakeholders, an issue came up with
files that had the same stem but different extensions and this caused
an issue copying files. See #36.

How this addresses that need:
* Change the way the NAS subfolder is named to use the full filename
and not just the stem
* Update the tests to reflect this new way of creating the "clean" NAS
subfolder name

Side effects of this change:
None.

Relevant ticket(s):
* https://mitlibraries.atlassian.net/browse/IR-214

Resolves #36